feat: MeetingVote 에서 contents 분리
