+++ src/pages/MeetingVote/MeetingVote.tsx
@@ -1,25 +1,22 @@
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
-import { UIEvent, useEffect, useEffect, useRef, useState } from 'react';
+import { UIEvent, useEffect, useRef, useState } from 'react';
 import { useResizeDetector } from 'react-resize-detector';
 import { useNavigate, useSearchParams } from 'react-router-dom';
-import { useRecoilState, useRecoilValue, useResetRecoilState, useSetRecoilState } from 'recoil';
+import { useRecoilValue, useSetRecoilState } from 'recoil';
 
 import WritingHands from '../../assets/writing.svg';
 import { ScrollDownFloatingButton } from '../../components/buttons/ScrollDownFloatingButton';
 import { Loading } from '../../components/Loading';
-import { Contents, Header, HeaderContainer, Page } from '../../components/pageLayout';
+import { Header, HeaderContainer, Page } from '../../components/pageLayout';
 import { FlexVertical } from '../../components/styled';
-import { UserList, UserListData } from '../../components/UserList/UserList';
-import { VoteTable } from '../../components/VoteTable/VoteTable';
-import { MeetingType } from '../../constants/meeting';
 import { useMeeting } from '../../hooks/useMeeting';
-import { useMeetingViewVoteMode } from '../../hooks/useMeetingVote';
 import { useVotings } from '../../hooks/useVotings';
 import { currentUserStateFamily } from '../../stores/currentUser';
 import { currentUserVotingSlotsState } from '../../stores/currentUserVotingSlots';
-import { userListState, votingsState as votingRecoilState } from '../../stores/voting';
-import { PrimaryBold, VoteTableWrapper } from '../../templates/MeetingView/styled';
+import { votingsState as votingRecoilState } from '../../stores/voting';
+import { PrimaryBold } from '../../templates/MeetingView/styled';
+import MeetingVoteContents from '../../templates/MeetingVote/MeetingVoteContents';
 import MeetingVoteFooter from '../../templates/MeetingVote/MeetingVoteFooter';
 
 function MeetingVote() {
@@ -29,24 +26,14 @@ function MeetingVote() {
   const { votings, isFetching: isVotingsFetcing } = useVotings();
   const isFetching = isMeetingFetching && isVotingsFetcing;
 
-  const [currentUser, setCurrentUser] = useRecoilState(currentUserStateFamily(meetingId));
-  const resetCurrentUser = useResetRecoilState(currentUserStateFamily(meetingId));
+  const currentUser = useRecoilValue(currentUserStateFamily(meetingId));
   const setCurrentUserVotingSlots = useSetRecoilState(currentUserVotingSlotsState);
   const isNewUser = !currentUser;
 
-  const [votingsState, setVotingsState] = useRecoilState(votingRecoilState);
-  const userList = useRecoilValue(userListState);
-  const checkedUserList = userList.map((user) => ({
-    ...user,
-    checked: user.id === currentUser?.id,
-  }));
-
   const [hasMoreBottomScroll, setHasMoreBottomScroll] = useState<boolean>(false);
 
+  const setVotingsState = useSetRecoilState(votingRecoilState);
   const navigate = useNavigate();
-
-  const { voteTableDataList, handleClickVoteTableSlot, handleClickVoteTableDate } =
-    useMeetingViewVoteMode(meeting);
 
   useEffect(() => {
     const isFromSharedURL = searchParams.get('ref') === 'share';
@@ -75,29 +62,6 @@ function MeetingVote() {
     },
   });
 
-  const handleClickUser = (checked: boolean, clickedUser: UserListData) => {
-    if (!meeting) {
-      return;
-    }
-
-    const isCurrentUserClicked = currentUser?.id === clickedUser.id;
-    if (isCurrentUserClicked) {
-      setCurrentUser(undefined);
-      resetCurrentUser();
-      setCurrentUserVotingSlots([]);
-      return;
-    }
-
-    setCurrentUser({
-      id: clickedUser.id,
-      username: clickedUser.username,
-    });
-
-    const previousVoting = votingsState.find((voting) => voting.username === clickedUser.username);
-    const previousVotingSlots = previousVoting?.[meeting.type];
-    setCurrentUserVotingSlots(previousVotingSlots ?? []);
-  };
-
   const handlePageScroll = (event: UIEvent<HTMLDivElement>) => {
     const { scrollTop, scrollHeight, clientHeight } = event.currentTarget;
     setHasMoreBottomScroll(scrollTop + clientHeight < scrollHeight);
@@ -121,7 +85,7 @@ function MeetingVote() {
     return <Loading />;
   }
 
-  if (!meeting || !voteTableDataList) {
+  if (!meeting) {
     return null;
   }
 
@@ -166,24 +130,7 @@ function MeetingVote() {
           </FlexVertical>
         </HeaderContainer>
       </Header>
-      <Contents>
-        <UserList
-          className="user-list"
-          users={checkedUserList}
-          onClick={handleClickUser}
-          selectedTooltipText="새로운 유저로 투표하려면 다시 클릭해주세요"
-        >
-          <UserList.Title color="primary">투표 현황</UserList.Title>
-        </UserList>
-        <VoteTableWrapper>
-          <VoteTable
-            onDateClick={handleClickVoteTableDate}
-            onSlotClick={handleClickVoteTableSlot}
-            data={voteTableDataList}
-            headers={meeting.type === MeetingType.date ? ['투표 현황'] : ['점심', '저녁']}
-          />
-        </VoteTableWrapper>
-      </Contents>
+      <MeetingVoteContents />
       <ScrollDownFloatingButton onClick={handleScollDownButtonClick} show={hasMoreBottomScroll} />
       <MeetingVoteFooter />
     </Page>
